press u32

+++ src/api/lps22df_internal.rs
@@ -346,12 +346,12 @@ impl<B: super::BusOperation> super::Lps22df<B> {
         Ok(val)
     }
 
-    pub(super) fn press_out_xl_l_h_get(&mut self) -> Result<u32, Error<B::Error>> {
+    pub(super) fn press_out_xl_l_h_get(&mut self) -> Result<i32, Error<B::Error>> {
         let mut arr: [u8; 3] = [0; 3];
         self.read_from_register(Reg::PressOutXl, &mut arr)?;
         let val: u32 = arr[0] as u32 | (arr[1] as u32) << 8 | (arr[2] as u32) << 16;
 
-        Ok(val)
+        Ok(val as i32)
     }
 
     pub(super) fn press_out_h_get(&mut self) -> Result<u8, Error<B::Error>> {
